+++ dashtools/deploy/fileUtils.py
@@ -7,6 +7,7 @@ import os
 import re
 import subprocess
 from typing import Union
+from pipreqs import pipreqs
 
 
 def check_file_exists(root_path: os.PathLike, file_name: str) -> bool:
@@ -39,18 +40,19 @@ def create_requirements_txt(root_path: os.PathLike, destination: os.PathLike = N
         update: Optional boolean to update existing requirements.txt
     """
     print(
-        f'dashtools: {"Updating" if update else "Creating"} requirements.txt')
+        f'dashtools: {"Found requirements.txt. Updating" if update else "Creating requirements.txt"} to {root_path}...')
     try:
-        optional_path = f" --savepath {os.path.join(destination, 'requirements.txt')}" if destination else ""
+        args = {'<path>': root_path, '--encoding': 'utf8', '--pypi-server': None, '--proxy': None, '--proxy-auth': None,
+                '--proxy-type': None, '--requirement': None, '--requirements': None, '--version': None, '--use-local': None, '--savepath': None, '--diff': None, '--clean': None, '--print': None, '--force': None, '--mode': None}
+        args['--savepath'] = os.path.join(destination,
+                                          'requirements.txt') if destination else None
         if update:
-            subprocess.check_output(
-                f'pipreqs --encoding=utf8 --force {root_path}{optional_path}', shell=True)
+            args['--force'] = True
+            pipreqs.init(args)
         else:
-            subprocess.check_output(
-                f'pipreqs --encoding=utf8 {root_path}{optional_path}', shell=True)
-    except subprocess.CalledProcessError as e:
-        # pipreqs throws a SyntaxError if it encounters a non-ASCII character
-        # One reason may be that the user is not in a valid dash app directory
+            pipreqs.init(args)
+    except Exception as e:
+        # Catch any pipreqs exceiptions
         print(e)
         print('dashtools: Error creating requirements.txt')
         print('dashtools: Are you in a valid dash app directory?')